+++ studio_socket/Response.hpp
@@ -72,7 +72,7 @@ class Response
 		{
 			if (methods.size() == 0)
 				return true;
-			for (int i = 0; i < methods.size(); i++)
+			for (size_t i = 0; i < methods.size(); i++)
 			{
 				//std::cout << BLUE << (methods[i] == method) << RESET << std::endl;
 				if (methods[i] == method)
